requirements.txt
joblib==1.2.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
numpy==1.23.3
pandas==1.4.4
psycopg2==2.8.6
pytest==7.1.2
pytest-mock==3.7.0
scikit-learn==1.1.2
scipy==1.9.1
SQLAlchemy==1.4.32
Werkzeug==2.0.3